concord-common: extract GrepUtils from IOUtils
