+++ common/src/main/java/com/walmartlabs/concord/common/GrepUtils.java
@@ -1,0 +1,48 @@
+package com.walmartlabs.concord.common;
+
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2025 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import java.io.*;
+import java.util.ArrayList;
+import java.util.List;
+
+public final class GrepUtils {
+
+    public static List<String> grep(String pattern, byte[] ab) throws IOException {
+        return grep(pattern, new ByteArrayInputStream(ab));
+    }
+
+    public static List<String> grep(String pattern, InputStream in) throws IOException {
+        List<String> result = new ArrayList<>();
+        try (BufferedReader reader = new BufferedReader(new InputStreamReader(in))) {
+            String line;
+            while ((line = reader.readLine()) != null) {
+                if (line.matches(pattern)) {
+                    result.add(line);
+                }
+            }
+        }
+        return result;
+    }
+
+    private GrepUtils() {
+    }
+}
